Move in2out and out2in
